Refactor of ValidAddress trait
