+++ src/common/genesis_block.rs
@@ -64,7 +64,8 @@ impl Encode for GenesisBlock {
 #[cfg(test)]
 pub mod tests {
     use super::*;
-    use crate::common::address::{Address, ValidAddress};
+    use crate::common::address::Address;
+    use crate::traits::ValidAddress;
 
     use secp256k1::{RecoverableSignature, RecoveryId, Secp256k1};
