+++ src/common/signed_tx.rs
@@ -159,7 +159,7 @@ impl Valid for SignedTx {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::common::address::ValidAddress;
+    use crate::traits::ValidAddress;
     use rand::{thread_rng, Rng};
 
     #[test]
