+++ src/common/transaction.rs
@@ -1,6 +1,7 @@
 use std::error::Error;
 
-use crate::common::address::{Address, ValidAddress};
+use crate::common::address::Address;
+use crate::traits::ValidAddress;
 use crate::util::hash::hash;
 
 use secp256k1::{Error as SecpError, Message, RecoverableSignature, RecoveryId, Secp256k1};
